+++ src/main/java/com/app/shwe/repository/CategoryRepository.java
@@ -9,13 +9,13 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface CategoryRepository extends JpaRepository<Category, Long> {
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
 
     @Query(value = "SELECT * FROM category", nativeQuery = true)
     List<Category> findAllCategories();
 
     @Query(value = "SELECT * FROM category WHERE id = :id", nativeQuery = true)
-    Optional<Category> findCategoryById(@Param("id") Long id);
+    Optional<Category> findCategoryById(@Param("id") int id);
 
     @Modifying
     @Query(value = "INSERT INTO category (name, label) VALUES (:name, :label)", nativeQuery = true)
@@ -23,9 +23,9 @@ public interface CategoryRepository extends JpaRepository<Category, Long> {
 
     @Modifying
     @Query(value = "UPDATE category SET name = :name, label = :label WHERE id = :id", nativeQuery = true)
-    void updateCategory(@Param("id") Long id, @Param("name") String name, @Param("label") String label);
+    void updateCategory(@Param("id") int id, @Param("name") String name, @Param("label") String label);
 
     @Modifying
     @Query(value = "DELETE FROM category WHERE id = :id", nativeQuery = true)
-    void deleteCategoryById(@Param("id") Long id);
+    void deleteCategoryById(@Param("id") int id);
 }
