+++ src/main/java/com/app/shwe/repository/ReportRepository.java
@@ -1,13 +1,9 @@
 package com.app.shwe.repository;
 
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import com.app.shwe.model.Report;
 
-public interface ReportRepository extends JpaRepository<Report, Long> {
+public interface ReportRepository extends JpaRepository<Report, Integer> {
 
 }
